backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
guardrails-ai>=0.5.0
spacy>=3.7.0
pdfplumber>=0.10.0